Finished the 'about' page, including auto-biography, achievements, and UTAS unit results.
Started creating a contact page with a 'reveal email' button. To prevent scraping, email is stored using a very weak encryption (reversed), and decrypted when the user clicks on some blacked-out text.

Also changed the opacity of the white backgrounds to 0.8 (previously 0.6).

Also added a 'fonts.css' file for importing fonts arbitrarily.